Apply change: r7 | rmachet | 2009-02-26 12:03:11 -0800 (Thu, 26 Feb 2009) | 13 lines
-Fixed Makefiles in extensions.
-Created integrated build system that is based on $(EPICS_SITE_TOP)/RELEASE_SITE
which means the only thing that needs to be set is EPICS_SITE_TOP.
-Modified base to support environment variables in RELEASE files
THIS IS NOT TO BE PROPAGATED TO epics-tech, they do not want this hack.
I need to implement a more elegant solution.
-Created Makefile to build everything from the top without even defining
EPICS_SITE_TOP. At the top type 'make' and it will build everything
(usefull for the build machine mainly). FYI total build time is 22mn.


git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/base/current@2262 ececa60f-9f63-0410-97c7-ea05888eb260