Apply changes from: ------------------------------------------------------------------------ r396 | rmachet | 2009-05-21 10:33:52 -0700 (Thu, 21 May 2009) | 10 lines
WARNING:
Modified the default behavior of the build system: by default the build 
system will not build for the build machine architecture anymore, unless
you specify BUILD_FOR_HOST_ARCH=1 in configure/RELEASE (of a module).
This change was made because many modules do not support the 64bit architecture
as a running target.
This should not affect you if you build on a 32 bits machine because linux-x86 
is also in the list of targets


------------------------------------------------------------------------


git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/base/current@2271 ececa60f-9f63-0410-97c7-ea05888eb260