Merged in all PCDS makeBaseApp templates. All other LCLS and PCDS changes left out of this version.
RTEMS is not yet supported, only linux-x86 and linux-x86_64

This version does NOT include autosave and archive support
in base.  Instead, these features need to be supported
via custom PCDS config files in each module and IOC.
The advantage of this approach is that it will allow our modules
and IOC's to build against stock EPICS base distributions.

This version of the PCDS makeBaseApp templates generates
the required files, but I'll need to do more testing and
develop a script for pushing these changes to each of
our IOC's and modules before everyone can switch to 3.14.11.



git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/branches/merge/epics/base/current@4686 ececa60f-9f63-0410-97c7-ea05888eb260