Fix cpuEnv support in convertRelease.pl for CONFIG_SITE gnu-make functions.
This modifies the pattern for macro2 = $(macro1) variable assignments
to not try to expand macro1 if it contains spaces.
The most common example of this is gnu-make functions,
which follow the form $(functionName args).
Without this change patterns like this cause an endless loop in expandRelease().
CROSS_COMPILER_TARGET_ARCHS := $(filter-out vxWorks%,$(CROSS_COMPILER_TARGET_ARCHS))